A data sector number truncated above 32 MiB
absolute_data_sector took a uint16_t. Block 65536 is 32 MiB into the data
region, and a v6 volume is allowed a gibibyte. Every caller passes a 64-bit
block number taken from an extent, so the truncation happened silently at the
call site and the sector number wrapped: a read or a write past 32 MiB went to
a sector near the start of the region instead. No error, no short count, the
wrong bytes at the wrong place.

The parameter is now uint64_t, which is what every caller already had.

The other 16-bit block numbers in the file are left alone on purpose. They are
the legacy v3/v4/v5 on-disk node formats, where 16 bits is the format rather
than an oversight, and v5's 8192 data sectors sit well inside it.

This was not reachable by the v5 volumes anything currently boots, which is why
it sat here since v6 raised the ceiling. It is reachable now in the sense that
matters: a v6 volume can be grown into it by ordinary use.

tests/storage/test_xaiboot_fs_large_volume.c writes 140 files of 256 KiB --
35 MiB, past the boundary with room on either side -- and reads every one back
after all of them are written. The second pass is the one that matters: a
wrapped write lands on a sector belonging to a file written earlier, so the
damage is behind you when it happens and only a full re-read finds it. Built
against the previous code the read fails outright.

256 KiB per file because that is the largest single write the whole-file path
stages, so reaching 32 MiB takes files rather than one enormous one.

B-49.

+++ kernel/fs/xaiboot_fs.c
@@ -1308,8 +1308,18 @@ static xaios_status_t write_journal(xaios_xbfs_journal_t *journal) {
   return XAIOS_OK;
 }
 
-static uint64_t absolute_data_sector(uint16_t block_index) {
-  return active_data_start_sector() + (uint64_t)block_index;
+/* The block index is 64-bit because the volume is.
+ *
+ * This took a uint16_t, which stops at 65535 -- block 65536 is 32 MiB in, and
+ * a v6 volume is allowed a gibibyte. Every caller already passed a uint64_t
+ * taken from an extent, so the conversion happened silently at the call and
+ * the sector number wrapped: a read or a write past 32 MiB went to a sector
+ * near the start of the data region instead, hitting whatever was there. No
+ * error, no short count, the wrong bytes. Nothing reachable by the v5 volumes
+ * anything currently boots, which is why it sat here; v6 volumes are the ones
+ * that can grow into it. See B-49. */
+static uint64_t absolute_data_sector(uint64_t block_index) {
+  return active_data_start_sector() + block_index;
 }
 
 /* Claim enough blocks for a file, as few runs as possible.
